# api/openapi.yaml
openapi: 3.1.0
info:
  title: LXC File Hub API
  version: 1.0.0
  description: REST API for LXC File Hub (lxcfh) — section 10.2
servers:
  - url: /api/v1
    description: API v1 base path
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      operationId: listUsers
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Create user
      operationId: createUser
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userID}:
    parameters:
      - $ref: '#/components/parameters/userID'
    get:
      tags: [Users]
      summary: Get user
      operationId: getUser
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '204':
          description: Deleted
  /users/{userID}/password:
    parameters:
      - $ref: '#/components/parameters/userID'
    put:
      tags: [Users]
      summary: Change user password
      operationId: changePassword
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  format: password
      responses:
        '204':
          description: Password changed
  /users/{userID}/ssh-keys:
    parameters:
      - $ref: '#/components/parameters/userID'
    get:
      tags: [Users]
      summary: List user SSH keys
      operationId: listUserSSHKeys
      security:
        - sessionAuth: []
      responses:
        '200':
          description: SSH keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSSHKey'
    post:
      tags: [Users]
      summary: Add user SSH key
      operationId: addUserSSHKey
      security:
        - sessionAuth: []
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SSHKeyRequest'
      responses:
        '201':
          description: Added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSSHKey'
  /users/{userID}/ssh-keys/{keyID}:
    parameters:
      - $ref: '#/components/parameters/userID'
      - $ref: '#/components/parameters/keyID'
    delete:
      tags: [Users]
      summary: Delete user SSH key
      operationId: deleteUserSSHKey
      security:
        - sessionAuth: []
        - csrfToken: []
      responses:
        '204':
          description: Deleted
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: lxcfh_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    keyID:
      name: keyID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userID:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        display_name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
    UpdateUserRequest:
      type: object
      properties:
        display_name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator, viewer]
        enabled:
          type: boolean
    UserSSHKey:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        name:
          type: string
        fingerprint:
          type: string
        public_key:
          type: string
        created_at:
          type: string
          format: date-time
    SSHKeyRequest:
      type: object
      required: [name, public_key]
      properties:
        name:
          type: string
        public_key:
          type: string
...